refactor(e2e): split VAD data types into core and ROS headers
  * Create `data_types_core.hpp` and `data_types_ros.hpp` to separate ROS-free and ROS-specific types
  * Update VAD headers and design docs to reference the new split headers
  * Remove the old `data_types.hpp` header
