+++ e2e/autoware_tensorrt_vad/lib/data_types.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../src/data_types.hpp"
+#include "../src/data_types_ros.hpp"
 
 namespace autoware::tensorrt_vad
 {
